requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.3
joblib==1.3.2
firebase-admin==6.2.0
pyrebase4==4.7.1
python-dotenv==1.0.0
